[MIG] website_snippet_country_dropdown: Migration to version 18.0
TT54401
