+++ website_snippet_country_dropdown/README.rst
@@ -119,6 +119,7 @@ Contributors
    -  Alexandre Díaz
    -  Carlos Roca
    -  Ernesto Tejeda
+   -  Pilar Vargas
 
 Maintainers
 -----------
