+++ website_snippet_country_dropdown/demo/pages.xml
@@ -5,15 +5,15 @@
 <odoo>
     <template id="country_dropdown_demo_view" name="Country dropdown snippet test">
         <t t-call="website.layout">
-            <div id="wrap" class="oe_structure oe_empty">
+            <div id="wrap">
                 <section>
                     <div class="container readable mt32 mb32">
                         <div class="row">
                             <div class="col-md-12">
                                 <h1 class="text-center">This is a form</h1>
-                                <h3 class="text-muted text-center">
+                                <h4 class="text-muted text-center">
                                 Test here the country dropdown snippet behavior
-                            </h3>
+                            </h4>
                                 <form action="" class="form-horizontal">
                                     <div
                                         class="js_enabled"
