+++ website_snippet_country_dropdown/static/description/index.html
@@ -465,6 +465,7 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <li>Alexandre Díaz</li>
 <li>Carlos Roca</li>
 <li>Ernesto Tejeda</li>
+<li>Pilar Vargas</li>
 </ul>
 </li>
 </ul>
